requirements.txt
selenium
beautifulsoup4
requests
webdriver_manager
pytesseract
pillow
keras
tensorflow
matplotlib